requirements.txt
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.2
spacy==3.6.1
nltk==3.8.1
transformers==4.33.2
torch==2.0.1
python-dotenv==1.0.0
